+++ apps/web/app/(auth)/auth/signup-without-verification-success/page.tsx
@@ -1,4 +1,4 @@
-import BackToLoginButton from "@/components/auth/BackToLoginButton";
+import BackToLoginButton from "@/app/(auth)/auth/BackToLoginButton";
 import FormWrapper from "@/components/auth/FormWrapper";
 
 export default function SignupWithoutVerificationSuccess() {
